refactor(app): extract SessionService + CommandDispatcher (#77)
Behaviour-preserving move of two clusters out of Runtime in src/vanchor/app.py:

- runtime/sessions.py — SessionService: _build_blackbox, _install_blackbox_hook,
  blackbox_dumps, blackbox_path_for, create_backup, restore_backup, start_replay,
  stop_replay.
- runtime/commands.py — CommandDispatcher: record_command, command_audit,
  handle_command.

Runtime keeps thin shims (handle_command is THE command entry point for the
WS/REST server and the test suite, so its exact signature is preserved). Every
Runtime-owned reference inside the moved methods routes through self._rt.<m>() so
monkeypatch seams on rt keep working; handle_command's dispatch branches,
command-name strings, and argument handling are byte-for-byte identical.

Closes #77
